requirements.txt
streamlit
mysql-connector-python
pandas
google-generativeai
Pillow
pdf2image
python-dotenv
plotly
